apphub-service/apphub-service-skyxplore/apphub-service-skyxplore-data/src/test/java/com/github/saphyra/apphub/service/skyxplore/data/save_game/dao/production_order/ProductionOrderModelValidatorTest.java
package com.github.saphyra.apphub.service.skyxplore.data.save_game.dao.production_order;

import com.github.saphyra.apphub.api.skyxplore.model.game.ProductionOrderModel;
import com.github.saphyra.apphub.service.skyxplore.data.save_game.dao.GameItemValidator;
import com.github.saphyra.apphub.test.common.ExceptionValidator;
import org.junit.After;
import org.junit.Before;
import org.junit.Test;
import org.junit.runner.RunWith;
import org.mockito.InjectMocks;
import org.mockito.Mock;
import org.mockito.junit.MockitoJUnitRunner;

import java.util.UUID;

import static org.assertj.core.api.Assertions.catchThrowable;
import static org.mockito.BDDMockito.given;
import static org.mockito.Mockito.verify;

@RunWith(MockitoJUnitRunner.class)
public class ProductionOrderModelValidatorTest {
    @Mock
    private GameItemValidator gameItemValidator;

    @InjectMocks
    private ProductionOrderModelValidator underTest;

    @Mock
    private ProductionOrderModel model;

    @Before
    public void setUp() {
        given(model.getLocation()).willReturn(UUID.randomUUID());
        given(model.getExternalReference()).willReturn(UUID.randomUUID());
        given(model.getLocationType()).willReturn("location-type");
        given(model.getDataId()).willReturn("location-type");
        given(model.getAmount()).willReturn(2345);
        given(model.getRequiredWorkPoints()).willReturn(2345);
        given(model.getCurrentWorkPoints()).willReturn(2345);
    }

    @After
    public void validate() {
        verify(gameItemValidator).validate(model);
    }

    @Test
    public void nullLocation() {
        given(model.getLocation()).willReturn(null);

        Throwable ex = catchThrowable(() -> underTest.validate(model));

        ExceptionValidator.validateInvalidParam(ex, "location", "must not be null");
    }

    @Test
    public void nullLocationType() {
        given(model.getLocationType()).willReturn(null);

        Throwable ex = catchThrowable(() -> underTest.validate(model));

        ExceptionValidator.validateInvalidParam(ex, "locationType", "must not be null");
    }

    @Test
    public void nullExternalReference() {
        given(model.getExternalReference()).willReturn(null);

        Throwable ex = catchThrowable(() -> underTest.validate(model));

        ExceptionValidator.validateInvalidParam(ex, "externalReference", "must not be null");
    }

    @Test
    public void nullDataId() {
        given(model.getDataId()).willReturn(null);

        Throwable ex = catchThrowable(() -> underTest.validate(model));

        ExceptionValidator.validateInvalidParam(ex, "dataId", "must not be null");
    }

    @Test
    public void nullAmount() {
        given(model.getAmount()).willReturn(null);

        Throwable ex = catchThrowable(() -> underTest.validate(model));

        ExceptionValidator.validateInvalidParam(ex, "amount", "must not be null");
    }

    @Test
    public void nullRequiredWorkPoints() {
        given(model.getRequiredWorkPoints()).willReturn(null);

        Throwable ex = catchThrowable(() -> underTest.validate(model));

        ExceptionValidator.validateInvalidParam(ex, "requiredWorkPoints", "must not be null");
    }

    @Test
    public void nullCurrentWorkPoints() {
        given(model.getCurrentWorkPoints()).willReturn(null);

        Throwable ex = catchThrowable(() -> underTest.validate(model));

        ExceptionValidator.validateInvalidParam(ex, "currentWorkPoints", "must not be null");
    }

    @Test
    public void valid() {
        underTest.validate(model);
    }
}